% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provSummarizeR.R
\name{prov.summarize}
\alias{prov.summarize}
\title{prov.summarize}
\usage{
prov.summarize(create.zip = FALSE)
}
\arguments{
\item{create.zip}{if true all of the provenance data will be packaged up
into a zip file stored in the current working directory}
}
\description{
prov.summarize reads a JSON string that contains provenance and outputs
a text summary to the R console
}
