% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provSummarizeR.R
\name{prov.summarize.run}
\alias{prov.summarize.run}
\title{prov.summarize.run}
\usage{
prov.summarize.run(r.script, save = FALSE, create.zip = FALSE, ...)
}
\arguments{
\item{r.script}{the name of a file containing an R script}

\item{save}{if true saves the summary to the file prov-summary.txt on the 
provenance directory}

\item{create.zip}{if true all of the provenance data will be packaged up
into a zip file stored in the current working directory}

\item{...}{extra parameters are passed to the provenance collector}
}
\description{
prov.summarize.run executes a script, collects provenance, and outputs a
text summary to the console.
}
